Removes CSM container (for now) in favor of VSM
